admin/requirements.txt
customtkinter
GitPython
Pillow
